refactor: rename apy module to stats
